feat: support multiple Ledger app types (Official, Zondax) (#1242)
* feat: support multiple Ledger app types (Official, Zondax)

- encode the app kind in accounts.hw (0 software, 1 Zondax, 2 Official);
  NewAccount carries a single hw field
- introduce LedgerApp trait (replaces HWAPI) with kind-specific defaults;
  ZondaxApp wraps the legacy CLA 0x85 device code, StubLedger returns clear
  errors for software accounts, builds without the ledger feature and
  Official device operations (vk import, PCZT signing not implemented yet)
- Official accounts (transparent + ironwood pools) are created locally from
  the recovery seed via standard ZIP-32; pool masks are validated per app
  and pool mask literals are replaced by named POOL_* constants
- New Account form gains a Zondax/Official selector with per-app pool masks
  (T+S / T+Ironwood) synced into the form field
- drop obsolete Zondax-protocol device tests; add no-device creation tests
  for both app types

* fix(ledger): treat seed-based Official Ledger accounts as software

+++ rust/src/db.rs
@@ -888,7 +888,7 @@ pub const LEDGER_CODE: u32 = 1;
 pub async fn store_account_hw(
     connection: &mut SqliteConnection,
     account: u32,
-    hw_code: u32,
+    hw_code: u8,
     aindex: u32,
 ) -> Result<()> {
     sqlx::query("UPDATE accounts SET hw = ?2, aindex = ?3 WHERE id_account = ?1")
